requirements.txt
Pillow==8.4.0
Flask==3.1.0
tk
dotenv
google-api-python-client